requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
github3.py
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.2.2
jupyter-core==4.1.0
MarkupSafe==0.23
mistune==0.7.2
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.0
numpy==1.11.0
pandas==0.18.1
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
pyzmq==15.2.0
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.5
tornado==4.3
traitlets==4.2.1
sphinx==1.4.1
travis-sphinx
tinydb
